+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./humanBytes.js";
 export * from "./auditErrors.js";
 export * from "./scoring.js";
 export * from "./types.js";